Make android sample working (#8)
* Adapted headers for orx 1.12.
* Make android sample to work with Orx 1.12.
* Updated README and comments of gradle tasks.

+++ src/norx/joystick.nim
@@ -771,7 +771,7 @@ const
 const
   orxJOYSTICK_KU32_MIN_ID* = 1
 
-const orxJOYSTICK_KU32_MAX_ID* = (ord(orxJOYSTICK_BUTTON_NUMBER) / ord(orxJOYSTICK_BUTTON_SINGLE_NUMBER))
+const orxJOYSTICK_KU32_MAX_ID* = (ord(orxJOYSTICK_BUTTON_NUMBER) / ord(orxJOYSTICK_BUTTON_SINGLE_NUMBER)).toInt
 
 ## **************************************************************************
 ##  Functions directly implemented by orx core
